tree-option moved from tree-selection
