+++ src/lib/tree/public-api.ts
@@ -3,5 +3,6 @@ export * from './node';
 export * from './padding';
 export * from './toggle';
 export * from './tree-selection';
+export * from './tree-option';
 export * from './data-source/flat-data-source';
 export * from './data-source/nested-data-source';
